requirements.txt
llmlingua==0.2.2
chardet==5.2.0
langchain==0.2.12
langchain-community==0.2.11
langchain-core==0.2.29
langchain-openai==0.1.20
langchain-pinecone==0.1.3
langchain-text-splitters==0.2.2
langgraph==0.2.3
langgraph-checkpoint==1.0.2
langgraph-checkpoint-sqlite==1.0.0
langsmith==0.1.98
lark==1.1.9
openai==1.40.2
pillow==10.4.0
pinecone-client==5.0.1
PyMySQL==1.1.1
python-dotenv==1.0.1
PyYAML==6.0.2
pygraphviz==1.13
redis==5.1.1
pandas==2.2.3
langchain_groq==0.1.9
jsonschema==4.23.0
